rno1: wire psp e_paranthropus whole TU (5 fns) via shared header extraction from are
