+++ src/st/rno1/e_paranthropus.c
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "rno1.h"
+
+#ifdef VERSION_PSP
+// psp symbols keep the D_us_* name used by sibling asm objects
+#define g_EInitParanthropusThrownBone D_us_801806E8
+#endif
+
+#include "../e_paranthropus.h"
