--- src/st/rno1_psp/unk_1028.c
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../rno1/rno1.h"
-
-INCLUDE_ASM("st/rno1_psp/nonmatchings/rno1_psp/unk_1028", ParanthropusSetStep);
-
-INCLUDE_ASM("st/rno1_psp/nonmatchings/rno1_psp/unk_1028", EntityParanthropus);
-
-INCLUDE_ASM("st/rno1_psp/nonmatchings/rno1_psp/unk_1028", EntityParanthropusThrownBone);
-
-INCLUDE_ASM("st/rno1_psp/nonmatchings/rno1_psp/unk_1028", EntityParanthropusBoneHitbox);
-
-INCLUDE_ASM("st/rno1_psp/nonmatchings/rno1_psp/unk_1028", EntityParanthropusSkull);
